+++ app/models/renalware/diaverum/incoming/resolve_dialysate_flow_rate.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+require "attr_extras"
+
+module Renalware
+  module Diaverum
+    module Incoming
+      class ResolveDialysateFlowRate
+        pattr_initialize [:patient!, :patient_node!, :treatment_node!]
+
+        def call
+          treatment_node.DialysateFlow.presence ||
+            patient_node.current_dialysis_prescription&.DialysateFlow.presence ||
+            dialysate_flow_rate_from_hd_profile
+        end
+
+        private
+
+        def dialysate_flow_rate_from_hd_profile
+          patient.hd_profile&.document&.dialysis&.flow_rate
+        end
+      end
+    end
+  end
+end