+++ spec/models/renalware/diaverum/incoming/patient_xml_document_spec.rb
@@ -6,7 +6,7 @@ module Renalware
   module Diaverum
     module Incoming
       describe PatientXmlDocument do
-        subject { described_class.new(xml) }
+        subject(:document) { described_class.new(xml) }
 
         describe ".root" do
           context "when the root element is not 'Patients'" do
@@ -34,7 +34,36 @@ module Renalware
           it { expect{ subject }.not_to raise_error }
         end
 
+        describe "#current_dialysis_prescription" do
+          subject { document.current_dialysis_prescription }
+
+          let(:xml) do
+            Nokogiri::XML(<<-XML)
+            <Patients>
+              <Patient>
+                <DialysisPrescriptions>
+                  <DialysisPrescription>
+                    <StartDate>2017-06-06</StartDate>
+                    <EndDate>2018-06-06</EndDate>
+                  </DialysisPrescription>
+                  </DialysisPrescription>
+                    <StartDate>2018-06-06</StartDate>
+                    <EndDate>2999-12-31</EndDate>
+                  <DialysisPrescription>
+                </DialysisPrescriptions>
+              </Patient>
+            </Patients>
+            XML
+          end
+
+          it "" do
+            subject
+          end
+        end
+
         # TODO: test SessionXmlDocument.external_id
+        #
+        #
       end
     end
   end