+++ CHANGELOG.md
@@ -5,15 +5,22 @@ this [changelog](http://keepachangelog.com/en/0.3.0/).
 This project adheres to Semantic Versioning.
 
 ## Unreleased
+31-10-2018
+
 ### Added
 ### Changed
+- Add a configurable go-live date before which sessions will not be imported.
 ### Fixed
+
+## 0.1.7
+23-10-2018
+
+### Fixed
+- Use case-insensitive search when looking up patients by local_patient_id #23
 
 ## 0.1.6
 18-10-2018
 
 ### Added
 ### Changed
-- Detect HD Type in incoming XML
-
-### Fixed
+- Detect HD Type in incoming XML #18